LayoutTests/fast/mediastream/mediastream-gc.html

<!DOCTYPE HTML>
<html>
<body>
<script src="../../resources/gc.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video id="video" autoplay=""></video>
<script>
promise_test(async () => {
    if (!window.internals)
        return;

    for (let counter = 0; counter < 10; ++counter) {
        await navigator.mediaDevices.getUserMedia({audio:true, video : true});
        gc();
    }
    assert_less_than(internals.audioCaptureSourceCount(), 10); 
}, "Media streams can be gced if they are not expected to fire events - 1");

promise_test(async () => {
    if (!window.internals)
        return;

    for (let counter = 0; counter < 10; ++counter) {
        video.srcObject = await navigator.mediaDevices.getUserMedia({audio:true, video : true});
        video.srcObject.clone();
        await video.play();
        video.srcObject = null;
        gc();
    }
    assert_less_than(internals.audioCaptureSourceCount(), 20); 
}, "Media streams can be gced if they are not expected to fire events - 2");
</script>
</body>
</html>
